Renamed SolCmp to SolApplication Made SolApplication implement ApplicationListener Removed the SolAppListener, since it was just a wrapper for SolApplication, with no logic in it
