--- main/src/com/miloshpetrov/sol2/SolAppListener.java
@@ -1,38 +1,0 @@
-package com.miloshpetrov.sol2;
-
-import com.badlogic.gdx.ApplicationListener;
-
-public class SolAppListener implements ApplicationListener {
-
-  private SolCmp myCmp;
-
-  @Override
-  public void create() {
-    myCmp = new SolCmp();
-  }
-
-  @Override
-  public void resize(int width, int height) {
-    // ignore nagluho
-  }
-
-  @Override
-  public void render() {
-    myCmp.render();
-  }
-
-  @Override
-  public void pause() {
-    myCmp.paused();
-  }
-
-  @Override
-  public void resume() {
-    // ignore for now
-  }
-
-  @Override
-  public void dispose() {
-    myCmp.dispose();
-  }
-}
