+++ main/src/com/miloshpetrov/sol2/menu/OptionsScreen.java
@@ -2,7 +2,7 @@ package com.miloshpetrov.sol2.menu;
 
 import com.badlogic.gdx.*;
 import com.miloshpetrov.sol2.GameOptions;
-import com.miloshpetrov.sol2.SolCmp;
+import com.miloshpetrov.sol2.SolApplication;
 import com.miloshpetrov.sol2.ui.*;
 
 import java.util.ArrayList;
@@ -37,7 +37,7 @@ public class OptionsScreen implements SolUiScreen {
   }
 
   @Override
-  public void updateCustom(SolCmp cmp, SolInputManager.Ptr[] ptrs, boolean clickedOutside) {
+  public void updateCustom(SolApplication cmp, SolInputManager.Ptr[] ptrs, boolean clickedOutside) {
     SolInputManager im = cmp.getInputMan();
     MenuScreens screens = cmp.getMenuScreens();
     if (myResoCtrl.isJustOff()) {
@@ -58,17 +58,17 @@ public class OptionsScreen implements SolUiScreen {
   }
 
   @Override
-  public void drawBg(UiDrawer uiDrawer, SolCmp cmp) {
+  public void drawBg(UiDrawer uiDrawer, SolApplication cmp) {
 
   }
 
   @Override
-  public void drawImgs(UiDrawer uiDrawer, SolCmp cmp) {
+  public void drawImgs(UiDrawer uiDrawer, SolApplication cmp) {
 
   }
 
   @Override
-  public void drawText(UiDrawer uiDrawer, SolCmp cmp) {
+  public void drawText(UiDrawer uiDrawer, SolApplication cmp) {
   }
 
   @Override
@@ -82,12 +82,12 @@ public class OptionsScreen implements SolUiScreen {
   }
 
   @Override
-  public void onAdd(SolCmp cmp) {
+  public void onAdd(SolApplication cmp) {
 
   }
 
   @Override
-  public void blurCustom(SolCmp cmp) {
+  public void blurCustom(SolApplication cmp) {
 
   }
 }
