+++ main/src/com/miloshpetrov/sol2/ui/SolInputManager.java
@@ -87,7 +87,7 @@ public class SolInputManager {
 
   }
 
-  public void setScreen(SolCmp cmp, SolUiScreen screen) {
+  public void setScreen(SolApplication cmp, SolUiScreen screen) {
     for (int i = 0, myScreensSize = myScreens.size(); i < myScreensSize; i++) {
       SolUiScreen oldScreen = myScreens.get(i);
       removeScreen(oldScreen, cmp);
@@ -95,12 +95,12 @@ public class SolInputManager {
     addScreen(cmp, screen);
   }
 
-  public void addScreen(SolCmp cmp, SolUiScreen screen) {
+  public void addScreen(SolApplication cmp, SolUiScreen screen) {
     myToAdd.add(screen);
     screen.onAdd(cmp);
   }
 
-  private void removeScreen(SolUiScreen screen, SolCmp cmp) {
+  private void removeScreen(SolUiScreen screen, SolApplication cmp) {
     myToRemove.add(screen);
     List<SolUiControl> controls = screen.getControls();
     for (int i = 0, controlsSize = controls.size(); i < controlsSize; i++) {
@@ -120,7 +120,7 @@ public class SolInputManager {
     ptr.y = 1f * screenY / h;
   }
 
-  public void update(SolCmp cmp) {
+  public void update(SolApplication cmp) {
     boolean mobile = cmp.isMobile();
     if (!mobile) maybeFixMousePos();
 
@@ -202,7 +202,7 @@ public class SolInputManager {
     myToAdd.clear();
   }
 
-  private void updateCursor(SolCmp cmp) {
+  private void updateCursor(SolApplication cmp) {
     if (cmp.isMobile()) return;
     myMousePos.set(myPtrs[0].x, myPtrs[0].y);
     if (cmp.getOptions().controlType != GameOptions.CONTROL_KB) {
@@ -246,7 +246,7 @@ public class SolInputManager {
     }
   }
 
-  public void draw(UiDrawer uiDrawer, SolCmp cmp) {
+  public void draw(UiDrawer uiDrawer, SolApplication cmp) {
     for (int i = myScreens.size() - 1; i >= 0; i--) {
       SolUiScreen screen = myScreens.get(i);
 
@@ -289,11 +289,11 @@ public class SolInputManager {
     return myMouseOnUi;
   }
 
-  public void playHover(SolCmp cmp) {
+  public void playHover(SolApplication cmp) {
     myHoverSound.play(.7f * cmp.getOptions().volMul, .7f, 0);
   }
 
-  public void playClick(SolCmp cmp) {
+  public void playClick(SolApplication cmp) {
     myHoverSound.play(.7f * cmp.getOptions().volMul, .9f, 0);
   }
